Move socket code into a separate header file
We also don't use any Zephyr code in this file so technically we could also use it in the emulator to have real socket support!
